luks-home: unlock a luks-encrypted home partition using transient units
A user's luks-encrypted home must be a (GPT) partition with the label
given by the concatentation of the string 'home' with the md5 hash
string of the username (without '\n'). This exactly fits into the maximum
length of the GPT partition label.

The shell script has a convenience argument to set this encoded partition
label.

The script is to be executed by PAM (using pam_exec.so) on a login
attempt, which then tries to find the partition with given label
matching the encoded identifier.
Therefore usernames are not easily leaked in an offline state
(provided the filesystem containing the passwd files etc. is not
readable i.e. encrypted).
Transient systemd service and mount units are created if a matching
partition is found for the user being logged in with PAM.
The LUKS container is unlocked and the mapped decrypted device
is mounted. The transient units die, when all sessions of the
given user terminate and in the process unmount
and lock the LUKS paritition.

The password of the LUKS container must be the same as that of the
user account. More precisely, it must match the value of the token
exposed by pam_exec's 'expose_authtok', which is for example the
given sized chunk of the password.

This version of the script makes use of systemd-cryptsetup instead
of the standard cryptsetup command, to easily bind the encryption
of the LUKS container to values of PCRs of a Trusted Platform
Module. The script can also update PCRs by measuring the decryption
key of the LUKS container into the TPM, which allows to bind later
stages to the state of the TPM PCRs. It takes care to only update
the PCRs if they are in an expected state. This prevents the user
from locking later stages bound the updated TPM PCRs if they log
in and out more than once, which would naively mutate the PCRs
multiple times, leading to later stages bound to them not seeing an
expected value. The idea is that the user will want later stages
to require a specified user to have been logged in once, proving
they posses certain tokens (passwords, PINs or physical security
tokens).