luks-home.conf: an example configuration file
This configuration file specifies the location of the file containing
the expected value of the PCRs used as a prerequisiste condition for
extending the PCRs with the volume decryption key and a string
specifying which PCRs + banks are contained in it.
(The latter is technically not necessary and could be replaced by
parsing the json in the pcr-prerequisiste file.

The configuration file is to be placed at '/etc/luks-home.conf'